// src/app/shared/enums/content-moderation.enum.ts
export enum QuestionOrQuizIssueReportSeverity {
  High = 1,
  Low = 2,
  Medium = 3,
  UnderProcessing = 4,
}

export enum QuestionOrQuizIssueReportStatus {
  Accepted = 1,
  Ignore = 2,
  Pending = 3,
  UnderReview = 4,
}

export enum QuizRatingStatus {
  Accepted = 1,
  Ignore = 2,
  Pending = 3,
  UnderProcessing = 4,
}